// packages/canisters/src/dip721/dip721.did.d.ts
export interface InitArgs {
  'logo' : [] | [string],
  'name' : [] | [string],
  'custodians' : [] | [Array<Principal>],
  'symbol' : [] | [string],
}